Allow to map hashmap into a vector
Summary:
HashMap::mapToVector allows to build a vector from an hashmap.

In Rust, an iterator can be mapped into any type, the outcome
of the iteration is decided afterwards by a collect method.

In PHP, native functions like `array_map` preserves the initial type.
As such, HashMap::map returns an HashMap.

The mapToVector method allows to collect the mapped values into a vector.

Test Plan:
This is intended for Zed to be used with API string representation:

```lang=php
    $response = HashMap::from($data)
        ->mapToVector(
            fn($key, $value) => "$key: " . self::toString($value)
        )
        ->implode(", ");
```

Unit tests are provided for our the different callback shapes.

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D3214